Add flann dependency to fix kdtree, segmentation, and search
Previously, these libraries would appear to build, but would fail to
link when actually used in the deps of a cc_binary.

This commit also adds all the examples to the pcl build file. This will
help ensure the pcl libraries actually link properly in the future.

The examples that depend on visualization are added, but commented out
with a TODO.